% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.concurFLM.R
\name{coef.concurFLM}
\alias{coef.concurFLM}
\title{coef.concurFLM}
\usage{
\method{coef}{concurFLM}(object, t.new = NULL)
}
\arguments{
\item{object}{return object from \code{\link{vbvs_concurrent}},
\code{\link{vb_concurrent}}, or \code{\link{ols_concurrent}}}

\item{t.new}{vector indicating the desired coordinates on which the functional
  coefficients will be evaluated}
}
\value{
a data frame containing the evaluation points and coefficient function
  values on the indicated grid
}
\description{
Extract coefficient functions from a fitted concurrent functional linear model
}
\details{
This function is used to extract coefficient functions from a concurrent functional
linear model, possibly on a user-specified grid.
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

