% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cv_vbvs_concurrent.R
\name{cv_vbvs_concurrent}
\alias{cv_vbvs_concurrent}
\title{cv_vbvs_concurrent}
\usage{
cv_vbvs_concurrent(formula, id.var = NULL, data = NULL, Kt = 5, Kp = 2,
  v0 = seq(0.01, 0.1, 0.01), v1 = 100, SEED = 1, standardized = FALSE,
  t.min = NULL, t.max = NULL)
}
\arguments{
\item{Kt}{number of spline basis functions for coefficients and FPCs}

\item{Kp}{number of FPCs to estimate}

\item{v0}{tuning parameter vector; normal spike variance. defaults to 0.01 to 0.1 in increments of 0.01.}

\item{v1}{tuning parameter; normal slab variance}

\item{SEED}{seed value, used to ensure reproducibility of the training split}

\item{standardized}{logical; are covariates already standardized?}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). if `NULL`, taken to be minium observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). if `NULL`, taken to be minium observed value.}

\item{Y}{matrix of functional responses}

\item{X}{list of matrices, each of which contains the functional predictor for an individual subject}
}
\description{
Cross validation to choose the tuning parameter v0 in the variational bayes variable selection algorithm
for the linear functional concurrent model. Uses five-fold cross validation.
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

