% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols_concurrent.R
\name{ols_concurrent}
\alias{ols_concurrent}
\title{ols_concurrent}
\usage{
ols_concurrent(Y, X, Kt = 5)
}
\arguments{
\item{Y}{matrix of functional responses}

\item{X}{list of matrices, each of which contains the functional predictor for an individual subject}

\item{Kt}{number of spline basis functions for coefficients and FPCs}
}
\description{
Implements ordinary least squares for the linear functional concurrent model
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

