% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.flcm.R
\name{predict.flcm}
\alias{predict.flcm}
\title{predict.flcm}
\usage{
\method{predict}{flcm}(object, data.new = NULL, standardized = TRUE, ...)
}
\arguments{
\item{object}{return object from \code{\link{vbvs_concurrent}}, 
\code{\link{vb_concurrent}}, or \code{\link{ols_concurrent}}}

\item{data.new}{dataset for which fitted values should be computed. should be a dataframe with names the same
as those used to fit the model, including the time variable used to parameterize functional observations}

\item{standardized}{logical; are values in `data.new` standardized? if not, they will be standardized using the
same scheme as used in the original analysis}

\item{...}{these arguments are ignored}
}
\value{
a vector of fitted values
}
\description{
Compute predicted values for a new dataset based on a fitted concurrent functional linear model
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

