% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_variables.R
\name{standardize_variables}
\alias{standardize_variables}
\title{standardize_variables}
\usage{
standardize_variables(data.original, data.new = NULL, time.var, trmstrings,
  LHS = NULL)
}
\arguments{
\item{data.original}{data used to fit the model}

\item{data.new}{data to be standardized. by default, this is the same as 
\code{data.original}; however it may be different to allow predictions on new
datasets}

\item{time.var}{the variable name that indicates the time variable}

\item{trmstrings}{collection of variables to be standardized}

\item{LHS}{the response variable name}
}
\description{
This function standardizes predictor and response variables prior to fitting the
concurrent model using variable selection.
}

