% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vbvs_concurrent.R
\name{vbvs_concurrent}
\alias{vbvs_concurrent}
\title{vbvs_concurrent}
\usage{
vbvs_concurrent(formula, id.var = NULL, data = NULL, Kt = 5, Kp = 2,
  v0 = 0.01, v1 = 100, standardized = FALSE)
}
\arguments{
\item{formula}{formula for desired regression. should have form \code{ y ~ x1 + x2 + ... + x_k | t}, where \code{t} is the variable that parameterized observed functions}

\item{id.var}{variable giving subject ID vector}

\item{data}{optional data frame}

\item{Kt}{number of spline basis functions for coefficients and FPCs}

\item{Kp}{number of FPCs to estimate}

\item{v0}{tuning parameter; normal spike variance}

\item{v1}{tuning parameter; normal slab variance}

\item{standardized}{logical; are covariates already standardized?}
}
\description{
Implements variational bayes variable selection for the linear functional concurrent model
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

