% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vb_concurrent.R
\name{vb_concurrent}
\alias{vb_concurrent}
\title{vb_concurrent}
\usage{
vb_concurrent(Y, X, Kt = 5, Kp = 2)
}
\arguments{
\item{Y}{matrix of functional responses}

\item{X}{list of matrices, each of which contains the functional predictor for an individual subject}

\item{Kt}{number of spline basis functions for coefficients and FPCs}

\item{Kp}{number of FPCs to estimate}
}
\description{
Implements variational bayes for the linear functional concurrent model
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

