% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vbvs_concurrent.R
\name{vbvs_concurrent}
\alias{vbvs_concurrent}
\title{vbvs_concurrent}
\usage{
vbvs_concurrent(Y, X, Kt = 5, Kp = 2, v0 = 0.01, v1 = 100)
}
\arguments{
\item{Y}{matrix of functional responses}

\item{X}{list of matrices, each of which contains the functional predictor for an individual subject}

\item{Kt}{number of spline basis functions for coefficients and FPCs}

\item{Kp}{number of FPCs to estimate}

\item{v0}{tuning parameter; normal spike variance}

\item{v1}{tuning parameter; normal slab variance}
}
\description{
Implements variational bayes variable selection for the linear functional concurrent model
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

